+++ src/components/CardRegisterForm/components/ExpirationPeriodField/ExpirationPeriodField.stories.tsx
@@ -1,0 +1,22 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import ExpirationPeriodField from "./ExpirationPeriodField";
+
+const meta = {
+  title: "ExpirationPeriodField",
+  component: ExpirationPeriodField,
+} satisfies Meta<typeof ExpirationPeriodField>;
+
+export default meta;
+
+const expiredPeriodState = {
+  inputs: ["", ""],
+  onChange: () => {},
+};
+
+type Story = StoryObj<typeof ExpirationPeriodField>;
+
+export const Default: Story = {
+  args: {
+    expiredPeriodState,
+  },
+};
